removed the "driver/adc.h" and it broke most of the ntc code, need to look through it to see how i fd up

+++ main/ntc.h
@@ -3,7 +3,7 @@
 /*
  * Read the value from the NTC resistor.
  */
-double ntc_read(void);
+int ntc_read(void);
 
 /*
  * Initialise the ntc stuff
